+++ app/ix/core/plugins/ConsoleFilterPlugin.java
@@ -1,0 +1,7 @@
+package ix.core.plugins;
+
+/**
+ * Created by katzelda on 9/15/16.
+ */
+public class ConsoleFilterPlugin {
+}