+++ app/ix/core/util/Filters.java
@@ -1,0 +1,7 @@
+package ix.core.util;
+
+/**
+ * Created by katzelda on 9/15/16.
+ */
+public class Filters {
+}