+++ modules/ginas/test/ix/test/ix/test/server/SubstanceAPI.java
@@ -448,9 +448,15 @@ public class SubstanceAPI {
             return response.getStatus();
         }
         public boolean isValid(){
+            if(getHttpStatus() !=200){
+                return false;
+            }
+            if(response.getBody().isEmpty()){
+                return false;
+            }
             return SubstanceJsonUtil.isValid(asJson());
         }
-        
+
         public boolean isNull(){
             return SubstanceJsonUtil.isLiteralNull(asJson());
         }
@@ -458,6 +464,7 @@ public class SubstanceAPI {
         public JsonNode asJson(){
             if(js ==null) {
                 js= session.extractJSON(response);
+                System.out.println(js.toString());
             }
             return js;
         }